Almost full normed and optied

+++ cpp00/ex01/includes/Contact.hpp
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/16 13:33:23 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/02/04 15:09:14 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/02/04 16:01:36 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ class Contact
 		Contact& operator=(Contact const &rhs);
 
     	void	set_contact_info();
-    	void	display_phonebook(int index) const;
-		void	display_contact(int index) const;
+    	void	display_phonebook(uint8_t index) const;
+		void	display_contact(uint8_t index) const;
 		
 	private:
 		std::string	_first_name;
