+++ cpp01/ex01/srcs/main.cpp
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 18:54:17 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/02/04 14:23:46 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/02/04 15:50:58 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 int main()
 {
-    int horde_size = 5;
+    uint8_t horde_size = 5;
     std::string zombie_name = "Horde_Zombie";
 
     Zombie* zombie_array = zombie_horde(horde_size, zombie_name);
 
-    for (int i = 0; i < horde_size; ++i)
+    for (uint8_t i = 0; i < horde_size; ++i)
         zombie_array[i].announce();
     delete[] (zombie_array);
